draft_src/requirements.txt
# pin versions as appropriate for your environment
torch
transformers
accelerate
datasets
safetensors
trl
peft
bitsandbytes
sentencepiece
huggingface_hub
wandb
pandas
numpy
scikit-learn
pyyaml
tqdm
yfinance
